Fix build errors and add missing widgets
